app/django/requirements.txt
Django==1.11.11
gunicorn==19.7.1
psycopg2==2.7.1

django-debug-toolbar
